Add detection of libraries that will need to be updated
Adds detection for the libraries that will need to be migrated to
AndroidX. Upon detection they will be notified to the user about what
file they are in and what they should be updated to. The versions should
be somewhat recent, but with so many being in alpha, beta and rc stage
it's likely they will change faster than the script can keep up with.

Tests are deleted simply due to them being almost useless. Adding more
testing as well as integration tests and bench marking would be a
stretch for the project, but a good thing to have if we want to ever
release this publicly.