Greatly improve performance by limiting disk IO
After profiling the code, 95% of the execution time was spent on writing
every line of every file out to a temporary file. Since a majority of
files don't change and this is an expensive task with the potential for
polluting a directory with .tmp files, the lines and any changes are
simply buffered in memory instead and then only opening a temporary file
when a change will be made and we want to persist it to disk. This keeps
almost all logic and work entirely in memory which dropped almost 95% of
the execution time (~16s to ~0.95s).

Also included in this patch is catching two edge cases. First, for lint
baseline files, where the lint error occurs in an xml file causing the
usual '<' or '>' tag starts to be converted to '&lt;' and '&gt;'. Adding
the ';' character as a valid boundary before checking for mappings seems
to solve this. Though it hasn't been tested if the baseline will still
register correctly if the class name length changes. Second, for full
class path function parameters.